Set tabindex and allow null as permissible value.
Currently when `tabindex` property is set on the toggle button it isn't passed inside template so it always defaults to `0`.

Also, allowing `null` as permissible value for `tabindex` will enable user to make the toggle button un-focusable. This is required to avoid the focus styles being applied during toggling the button.